// MS_rosterfrontend/projects/client/src/lib/view-models/client-GuardianInfo.ts
export interface GuardianInfo {
    clientId?: number;
    gender?: number;
    salutation?: number;
    name?: string;
    relationShip: number;
    relationship: number;
    contactNo?: string;
    email?:string;
    emailId?:string;
    phoneNo?: string;
    mobileNo?: string;
    relationShipName?: string;
    lastName?: string;
    middleName?: string;
    firstName?:string;
    otherRelation?:string;
    genderName?:string;
    
}